refactor: wrap plugins and configs as bicep context
