+++ packages/fx-core/src/common/azure-hosting/hostingFactory.ts
@@ -4,14 +4,14 @@
 import { AzureHosting } from "./azureHosting";
 import { FunctionHosting } from "./functionHosting";
 import { BotHosting } from "./botServiceHosting";
-import { HostType } from "./interface";
+import { HostType } from "./interfaces";
 
 const HostingMap: { [key: string]: () => AzureHosting } = {
   [HostType.Function]: () => new FunctionHosting(),
   [HostType.BotService]: () => new BotHosting(),
 };
 
-export class HostingResourceFactory {
+export class AzureHostingFactory {
   static createHosting(hostType: HostType): AzureHosting {
     if (HostingMap[hostType] !== undefined) {
       return HostingMap[hostType]();
